Fix typos and correct SplitMix32 result_type
Corrects several comment typos ("signle"→"single", "gave"→"have", "accidentaly"→"accidentally") and changes SplitMix32::result_type from std::uint64_t to std::uint32_t to match the intended 32-bit generator behavior, improving clarity and correctness.

+++ random.h
@@ -33,13 +33,13 @@ template <class ResultType>
   state = (state ^ (state >> 27)) * 0x94D049BB133111EB;
   return static_cast<ResultType>(state ^ (state >> 31));
   // some of the 16/32-bit PRNGs have bad correlation on the successive seeds, this usually
-  // can be alleviated by using a signle iteration of a "good" PRNG to pre-mix the seed
+  // can be alleviated by using a single iteration of a "good" PRNG to pre-mix the seed
 }
 
 template <class T>
 constexpr T _default_seed = (std::numeric_limits<T>::max() / 2) + 1;
-// an "overall decent" default seed - doesn't gave too many zeroes,
-// unlikely to accidentaly match with a user-defined seed
+// an "overall decent" default seed - doesn't have too many zeroes,
+// unlikely to accidentally match with a user-defined seed
 
 namespace generators {
 
@@ -60,7 +60,7 @@ namespace generators {
 //
 class SplitMix32 {
  public:
-  using result_type = std::uint64_t;
+  using result_type = std::uint32_t;
 
  private:
   result_type s{};
